Add raw and processed data for entanglement window analysis
Several components to this commit. First one is to rename the raw data and move it from inst/extdata to data-raw folder.

Second is to add in the code to process that raw data, and make useable rda files within the package. The script to do this is prepAmyEntanglementData.R

Third is to add the documentation for the new rda files created for subsequent analysis.