Update health calculation to place NAs after median death estimate
I'd been thinking that the way I'd calculated health to zero out times before firstSight, but not 0 out times after estimated death was inconsistent. This changes that to find the median estimate of death using the getDeaths() function, and then use that to place NAs for health after that time. Then it gets sent to the function to make the health anomaly.