Alter makeEvents() to account for the actual known death data from deadTable
What this means is that when I first wrote makeEvents() I was just asking for a logical about whether the animal was estimated to be dead with ng == ng before dcut. But this was incomplete, because I actually know the deaths. So I modified the test in makeEvents() to link up to data in deadTable for calculating the known deaths. I also had to update the help files, as well as the raw data.