Change flag for known dead animals
What I had in here wasn't working in the sense that the logic for the knownD animals was correct, but was getting over written by the subsequent if statements. So I've changed the order to make sure this variable is done last. This came up because in the survival data, I was only getting three animals that were known dead.