Add boxplot labels above the "Unimpacted" categories
The way I calculated these values was to go back to the functions that prep the data for the plotting. In that code, there are two data frames that make up the health anomaly values that go into the two left-hand most boxes

I calculated the number of animals in each that had at least one positive value, e.g. dim(repHealth)[1] - sum(apply(repHealth, 1, function(x) all(is.na(x)))), where that logic says find the number of rows in repHealth, and subtract from that the number of whales (rows) that are all na.