Correct the Survival/Entanglement Start Times
I had been extracting the start of the window based on the deathyr, but in reality I only wanted to extract the end from that. The beginning of the comparison window is to come from the end of the entanglement window.

I've also added in a bit more text to explain the next steps, i.e. processing the differences.