Add the Code to Add in the Reference Animals Who Have Never Been Entangled
The first part of this loop found the animals that have experienced an entanglement, but who aren't currently entangled.

The second part of the loop - added with this commit - finds the additional reference animals who have never been entangled and who are alive at the same time as the entangled individual.