Fix the Bug of Inexding Over EGNo
This is an important change. I had been incorrectly looping over EGNo and then using the position of the whale in that vector to update fate. But that i vector was updating fate in the wrong places. This fixes it by establishing a variable called anidx, which is linked off ID -- not off length(unique(egno))