Correct the Indexing on the Cumulative Numbers of Animals that Are Entangled
Note that this has had to change because I've added in the dead category. So it's not 100% satisfactory as a result. Specifically, towards the end you start to see the numbers of minor entanglements drop, which is because of deaths.

One possible way to address that would be to make a whole separate fate matrix, where one only includes entangled fate, and one includes death fates as well as the entangled ones.