Change the Indexing for the Events that Lack a Start Date
When trying to figure out how many animals were lacking a start date, I unpacked this indexing a little bit more carefully, and found that the seq_along() was incorrect. The later matching is correct, i.e. the matching within the loop, but the indexing is wrong, which means the goal of the indexing, to update the empty start dates for calves and juveniles, is not being done properly. I only ended up fixing 25 animals, instead of the 130 that are in the calf/juvenile data frame