Change the Input Data to Be from a Previously Saved Entanglement Array; Keep No Start-Date Data
This is one that I had created on my old mac at St Andrews, which is now dead. So I have to bring in the data that I'd previously created. Note that this will change to make it more reproducible.

Also, another big change is that here I keep the data/events that do not have a start date. For these I simply say the start date is 3 months prior to the end date.